+++ package_documentation.py
@@ -11,8 +11,6 @@ html_path_external = pathlib.Path("build/docs/html").resolve()
 html_path_internal = pathlib.Path("build/docs/html-internal").resolve()
 html_path = html_path_external if html_path_external.exists() else html_path_internal
 # TODO: figure out the setuptools and setuptools_scm solution for including data files
-modsim_template_path = pathlib.Path("modsim_template").resolve()
-tutorials_path = pathlib.Path("tutorials").resolve()
 # FIXME: setuptools claims that these should be included by default, but they aren't
 readme_path = pathlib.Path("README.rst").resolve()
 pyproject = pathlib.Path("pyproject.toml").resolve()
@@ -21,8 +19,6 @@ new_paths = [
     (prefix / "share/man/man1", man_path),
     (prefix / "man/man1", man_path),
     (sp_dir / pkg_name / "docs", html_path),
-    (sp_dir / pkg_name / "modsim_template", modsim_template_path),
-    (sp_dir / pkg_name / "tutorials", tutorials_path),
     (sp_dir / pkg_name / "README.rst", readme_path),
     (sp_dir / pkg_name / "pyproject.toml", pyproject),
 ]