+++ docs/changelog.rst
@@ -8,6 +8,14 @@ Changelog
 0.7.5 (unreleased)
 ******************
 
+New Features
+============
+- Add experimental builder support for Quinoa (:issue:`550`, :merge:`676`). By `Kyle Brindley`_.
+
+Documentation
+=============
+- Add work-in-progress tutorial for Quinoa (:issue:`550`, :merge:`676`). By `Kyle Brindley`_.
+
 ******************
 0.7.4 (2023-10-26)
 ******************